+++ lib/src/throughput/HarmonicMeanEstimator.ts
@@ -1,4 +1,5 @@
 import type { ResourceTiming } from '../request/ResourceTiming.js';
+import { getBandwidthBps } from '../utils/getBandwidthBps.js';
 import type { ThroughputEstimator } from './ThroughputEstimator.js';
 
 /**
@@ -19,11 +20,7 @@ export class HarmonicMeanEstimator implements ThroughputEstimator {
 		let value: number = 0;
 
 		for (let i = 0; i < this.samples.length; i++) {
-			const sample = this.samples[i];
-			const durationSeconds = sample.duration / 1000;
-			const bandwidthBps = sample.encodedBodySize * 8 / durationSeconds;
-
-			value += 1 / bandwidthBps;
+			value += 1 / getBandwidthBps(this.samples[i]);
 		}
 
 		return this.samples.length / value;